requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-heroku==0.3.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
newsapi-python==0.2.5
psycopg2==2.8.3
pylint==2.3.1
pytz==2019.2
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.24.3
whitenoise==4.1.3
wrapt==1.11.2
gunicorn==19.7.1
